Extensions: bring files in correct order
