+++ src/App/GeoFeatureGroupExtensionPyImp.cpp
@@ -24,27 +24,27 @@
 
 #include "PreCompiled.h"
 
-#include "App/GeoFeatureGroup.h"
+#include "App/GeoFeatureGroupExtension.h"
 
-// inclusion of the generated files (generated out of GeoFeatureGroupPy.xml)
-#include "GeoFeatureGroupPy.h"
-#include "GeoFeatureGroupPy.cpp"
+// inclusion of the generated files (generated out of GeoFeatureGroupExtensionPy.xml)
+#include "GeoFeatureGroupExtensionPy.h"
+#include "GeoFeatureGroupExtensionPy.cpp"
 
 using namespace App;
 
 // returns a string which represents the object e.g. when printed in python
-std::string GeoFeatureGroupPy::representation(void) const
+std::string GeoFeatureGroupExtensionPy::representation(void) const
 {
     return std::string("<GeoFeatureGroup object>");
 }
 
 
-PyObject *GeoFeatureGroupPy::getCustomAttributes(const char* /*attr*/) const
+PyObject *GeoFeatureGroupExtensionPy::getCustomAttributes(const char* /*attr*/) const
 {
     return 0;
 }
 
-int GeoFeatureGroupPy::setCustomAttributes(const char* /*attr*/, PyObject* /*obj*/)
+int GeoFeatureGroupExtensionPy::setCustomAttributes(const char* /*attr*/, PyObject* /*obj*/)
 {
     return 0;
 }
