+++ src/App/OriginGroupExtension.h
@@ -23,7 +23,7 @@
 #ifndef ORIGINGROUP_H_QHTU73IF
 #define ORIGINGROUP_H_QHTU73IF
 
-#include "GeoFeatureGroup.h"
+#include "GeoFeatureGroupExtension.h"
 #include "PropertyLinks.h"
 
 namespace App {
