+++ src/App/OriginGroupExtensionPy.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <GenerateModel xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="generateMetaModel_Module.xsd">
   <PythonExport
-      Father="GeoFeatureGroupPy"
-      Name="OriginGroupPy"
-      Twin="OriginGroup"
-      TwinPointer="OriginGroup"
-      Include="App/OriginGroup.h"
+      Father="GeoFeatureGroupExtensionPy"
+      Name="OriginGroupExtensionPy"
+      Twin="OriginGroupExtension"
+      TwinPointer="OriginGroupExtension"
+      Include="App/OriginGroupExtension.h"
       Namespace="App"
-      FatherInclude="App/GeoFeatureGroupPy.h"
+      FatherInclude="App/GeoFeatureGroupExtensionPy.h"
       FatherNamespace="App">
     <Documentation>
       <Author Licence="LGPL" Name="Alexander Golubev" EMail="fatzer2@gmail.com" />
