+++ src/App/OriginGroupExtensionPyImp.cpp
@@ -23,16 +23,16 @@
 
 #include "PreCompiled.h"
 
-#include "App/OriginGroup.h"
+#include "App/OriginGroupExtension.h"
 
 // inclusion of the generated files (generated out of OriginGroupPy.xml)
-#include "OriginGroupPy.h"
-#include "OriginGroupPy.cpp"
+#include "OriginGroupExtensionPy.h"
+#include "OriginGroupExtensionPy.cpp"
 
 using namespace App;
 
 // returns a string which represents the object e.g. when printed in python
-std::string OriginGroupPy::representation(void) const
+std::string OriginGroupExtensionPy::representation(void) const
 {
     return std::string("<OriginGroup object>");
 }
@@ -43,12 +43,12 @@ std::string OriginGroupPy::representation(void) const
 
 
 
-PyObject *OriginGroupPy::getCustomAttributes(const char* /*attr*/) const
+PyObject *OriginGroupExtensionPy::getCustomAttributes(const char* /*attr*/) const
 {
     return 0;
 }
 
-int OriginGroupPy::setCustomAttributes(const char* /*attr*/, PyObject* /*obj*/)
+int OriginGroupExtensionPy::setCustomAttributes(const char* /*attr*/, PyObject* /*obj*/)
 {
     return 0; 
 }
